pyproject.toml
[project]
name = "data-quality-assessment-and-preprocessing"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.3",
    "pytest>=9.0.3",
    "pyyaml>=6.0.3",
]
